requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt setup.py
#
apted==1.0.3
    # via deepdoctection (setup.py)
attrs==21.4.0
    # via jsonlines
catalogue==2.0.7
    # via deepdoctection (setup.py)
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
distance==0.1.3
    # via deepdoctection (setup.py)
filelock==3.6.0
    # via huggingface-hub
fonttools==4.37.3
    # via matplotlib
huggingface-hub==0.4.0
    # via deepdoctection (setup.py)
idna==3.3
    # via requests
importlib-metadata==4.11.2
    # via deepdoctection (setup.py)
jsonlines==3.0.0
    # via deepdoctection (setup.py)
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.1
    # via deepdoctection (setup.py)
lxml-stubs==0.4.0
    # via deepdoctection (setup.py)
matplotlib==3.6.0
    # via pycocotools
mock==4.0.3
    # via deepdoctection (setup.py)
networkx==2.7.1
    # via deepdoctection (setup.py)
numpy==1.22.3
    # via
    #   contourpy
    #   deepdoctection (setup.py)
    #   matplotlib
    #   opencv-python
    #   pycocotools
opencv-python==4.5.4.60
    # via deepdoctection (setup.py)
packaging==21.3
    # via
    #   deepdoctection (setup.py)
    #   huggingface-hub
    #   matplotlib
pillow==9.2.0
    # via matplotlib
pycocotools==2.0.5
    # via deepdoctection (setup.py)
pyparsing==3.0.7
    # via
    #   matplotlib
    #   packaging
pypdf2==1.27.9
    # via deepdoctection (setup.py)
python-dateutil==2.8.2
    # via matplotlib
pyyaml==6.0
    # via
    #   deepdoctection (setup.py)
    #   huggingface-hub
pyzmq==24.0.1
    # via deepdoctection (setup.py)
requests==2.27.1
    # via huggingface-hub
six==1.16.0
    # via python-dateutil
tabulate==0.8.10
    # via deepdoctection (setup.py)
termcolor==2.0.1
    # via deepdoctection (setup.py)
tqdm==4.63.0
    # via huggingface-hub
types-pyyaml==6.0.8
    # via deepdoctection (setup.py)
types-tabulate==0.8.9
    # via deepdoctection (setup.py)
types-termcolor==1.1.3
    # via deepdoctection (setup.py)
types-tqdm==4.64.6
    # via deepdoctection (setup.py)
typing-extensions==4.1.1
    # via huggingface-hub
urllib3==1.26.8
    # via requests
zipp==3.7.0
    # via importlib-metadata
python-dotenv==0.21.0
google-cloud-vision==3.2.0
fastapi==0.87.0